agent/etrendo-agent/requirements.txt
pydantic
google-cloud-aiplatform[adk]
google-cloud-bigquery
pyyaml
fastapi
uvicorn